+++ src/main/java/controller/Controller.java
@@ -1,6 +1,7 @@
 package controller;
 
-import model.*;
+import demain.*;
+import domain.*;
 import view.InputView;
 import view.OutputView;
