+++ src/main/java/demain/Card.java
@@ -1,4 +1,4 @@
-package model;
+package demain;
 
 public class Card {
     String card;
