+++ src/main/java/demain/ParticipantGameResult.java
@@ -1,4 +1,4 @@
-package model;
+package demain;
 
 import java.util.Map;
 import java.util.Set;
