+++ src/test/java/domain/CardNumberTest.java
@@ -1,5 +1,6 @@
-package model;
+package domain;
 
+import demain.CardNumber;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
