+++ src/test/java/domain/CardTest.java
@@ -1,5 +1,6 @@
-package model;
+package domain;
 
+import demain.Card;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
