+++ src/test/java/domain/CardTypeTest.java
@@ -1,5 +1,6 @@
-package model;
+package domain;
 
+import demain.CardType;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
