+++ src/test/java/domain/ParticipantGameResultTest.java
@@ -1,5 +1,6 @@
-package model;
+package domain;
 
+import demain.*;
 import org.junit.jupiter.api.Test;
 
 import java.util.*;
