+++ src/test/java/domain/RandomCardCreatorTest.java
@@ -1,5 +1,7 @@
-package model;
+package domain;
 
+import demain.Card;
+import demain.RandomCardCreator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
